Added adding and subtracting from account

+++ CashFlow/Services/BankAccountServices/IBankAccountService.cs
@@ -11,9 +11,9 @@ public interface IBankAccountService
     Task<ServiceResponse<GetBankAccountDto>> CreateBankAccount(AddBankAccountDto addBankAccountDto);
     Task<ServiceResponse<GetBankAccountDto>> UpdateBankAccount(UpdateBankAccountDto updateBankAccountDto);
     Task<ServiceResponse<List<GetBankAccountDto>>> DeleteBankAccount(int id);
-    Task<ServiceResponse<GetBankAccountDto>> AddBalance(int id, decimal amount);
-    Task<ServiceResponse<GetBankAccountDto>> SubtractBalance(int id, decimal amount);
-    Task<ServiceResponse<GetBankAccountDto>> TransferBalance(int id, int targetId, decimal amount);
-    Task<ServiceResponse<GetBankAccountDto>> AddCredit(int id, decimal amount);
-    Task<ServiceResponse<GetBankAccountDto>> SubtractCredit(int id, decimal amount); // Move money from balance to credit (deleting credit)
+    Task<ServiceResponse<GetBankAccountDto>> AddBalance(int id, double amount);
+    Task<ServiceResponse<GetBankAccountDto>> SubtractBalance(int id, double amount);
+    Task<ServiceResponse<GetBankAccountDto>> TransferBalance(int id, int targetId, double amount);
+    Task<ServiceResponse<GetBankAccountDto>> AddCredit(int id, double amount);
+    Task<ServiceResponse<GetBankAccountDto>> SubtractCredit(int id, double amount); // Move money from balance to credit (deleting credit)
 }
